Begin revamp of update script.
This is a first pass through the update script, and which begins to
add some of the features requested in #71. In particular, toggles are
added so that each individual part of the update script can be turned
off. As an example, a `--noParts` flag can be given to skip the flags
assignment. This can be used to bypass the parts issues mentioned in #67.