Change in search spelling suggestions; closes #109
Thanks to Raymond Cheng

This change does 3 small things:

1. The logic is changed to showing the alternative whenever there is an
alternative.

2. The wording is changed so as to make it a suggestion and not
something that implies there is something wrong with the spelling.

3. We've added some alternatives dealing with tricky issues such as
having multiple spellings of fibre by suggesting the user tries a search
with a wildcard. This leaves it up to the user to decide what is best
for her or him.