Logic of logging inactive tags
If tag.active is false, then the tag was already inactive and hence we
shouldn't tell the user that it became inactive on this run of the
update script. This change only affects logging and not the
functionality of the script.

Moreover, it seems to me that in the function checkInactivity() we
alsmost never need to call tag.save() because we only need to call it
when tag.active changes from True to False or vice versa.

Finally, it appears that with the current approach the tags that become
inactive remain in the database and their corresponding pages on the
server are still being served. I seem to recall that we intend to tell
visitors to these pages what/why/how, etc. I will make an issue.