Add a bit of space around equations
Very likely this isn't the ``correct'' solution to the problem I
describe below and in that case I would be happy for this commit to be
reverted. But I want the state of the repository to reflect what is
currently on the live server so we don't get any more confused than we
already are...

Because of the recent change in the css regarding equations we lost some
of the space around the equations and they looked rather scrunched. This
wasn't happening to the equations having labels (see below). My
solution: add some vertical margings.

The difference in treatement between the equations with labels and
without labels was due to a small ``mistake'' in plastex output
procedure for these: in one case the equation div was surrounded by a
<p>...</p> thing and in the other case it wasn't. I made the change in
plastex so in the future we don't get confused about this...